Add files via upload
A Video club that awards points to its customers based on the number
of videos purchased each month. The points are awarded as follows:
• If a customer purchases 0 dvd’s, the earning is 0 points.
• If a customer purchases 1 dvd, the earning is 2 points.
• If a customer purchases 2 dvd’s,the earning is 5 points.
• If a customer purchases 3 dvd’s, the earning is 10 points.
• If a customer purchases 4 or more dvd’s, the earning is 25 points.
Write a program that asks the user to enter the number of videos that
has been purchased for a month, then displays the number of points
awarded.

Write a program that calculates the BMI ( Body Mass Index ) for a user.
Use the "math" module's pow method in the program appropriately in
the formula.
Let the program ask the user for needed input:
• the weight ( in kg )
• the height ( in meter )
Then, display the BMI value on the screen with a message.
The message depends on the BMI index, if the person is:
• underweight
• normal weight - overweight
or
• has obesity
BMI index
Under weight: < 18.5
Normal weight: 18.5 - 24.99
Over weight: 25.0 - 29.99
Obesity: BMI of 30 or greater
Formula: BMI = weight / ( height 2 )

Property Tax
A county collects property taxes on the assessment value of property,
which is 60 percent of the property’s actual value.
For example, if an acre of land is valued at $10,000, its assessment value
is $6,000. The property tax is then 72¢ for each $100 of the assessment
value. The tax for the acre assessed at $6,000 will be $43.20. Write a
program that asks for the actual value of a piece of property and
displays the assessment value and property tax.


Write a func0on named evaluate_max that accepts two number
values as arguments and returns the value that is the greater of the two.
For example, if 12 and 19 are passed as arguments to the func0on, the
func0on should return 19. Ask the user to enter two values, the
program should then display the value that is the greater of the two.


Write a program that asks the user to enter five test scores. The program
should display a leaer grade for each score and the average test score.
Write the following func0ons in the program:
• calc_average This func0on should accept five test scores as
arguments and return the average of the scores.
• determine_grade This func0on should accept a test score as
an argument and return a leaer grade for the score based on the
following grading scale: Score Leaer Grade
90-100 A
80-89 B
70-79 C
60-69 D
Below 60 F